+++ src/main/java/br/dev/mhc/financialassistantapi/entities/accounts/BankAccount.java
@@ -1,5 +1,7 @@
 package br.dev.mhc.financialassistantapi.entities.accounts;
 
+import java.math.BigDecimal;
+
 import javax.persistence.Entity;
 import javax.persistence.Table;
 import javax.validation.constraints.NotNull;
@@ -17,36 +19,36 @@ public class BankAccount extends Account {
 
 	@NotNull(message = "Bank interest rate is a required field")
 	@PositiveOrZero(message = "The bank interest rate must be greater than or equal to zero")
-	private Double bankInterestRate;
+	private BigDecimal bankInterestRate;
 
 	@NotNull(message = "Limit value is a required field")
 	@PositiveOrZero(message = "The limit value must be greater than or equal to zero")
-	private Double limitValueBankAccount;
+	private BigDecimal limitValueBankAccount;
 
 	public BankAccount() {
 		this.setAccountType(AccountType.BANK_ACCOUNT);
 	}
 
-	public BankAccount(Long id, String name, Double balance, Double bankInterestRate, Double limitValueBankAccount,
+	public BankAccount(Long id, String name, BigDecimal balance, BigDecimal bankInterestRate, BigDecimal limitValueBankAccount,
 			User user) {
 		super(id, name, balance, AccountType.BANK_ACCOUNT, user);
 		this.bankInterestRate = bankInterestRate;
 		this.limitValueBankAccount = limitValueBankAccount;
 	}
 
-	public Double getBankInterestRate() {
+	public BigDecimal getBankInterestRate() {
 		return bankInterestRate;
 	}
 
-	public void setBankInterestRate(Double bankInterestRate) {
+	public void setBankInterestRate(BigDecimal bankInterestRate) {
 		this.bankInterestRate = bankInterestRate;
 	}
 
-	public Double getLimitValueBankAccount() {
+	public BigDecimal getLimitValueBankAccount() {
 		return limitValueBankAccount;
 	}
 
-	public void setLimitValue(Double limitValueBankAccount) {
+	public void setLimitValue(BigDecimal limitValueBankAccount) {
 		this.limitValueBankAccount = limitValueBankAccount;
 	}
 }
